Add a generic drivetrain subsystem
This subsystem will be instatiated in all future robot codes. To adapt
to different drivetrains, a DrivetrainConfig struct will be customized.

This subsystem includes both an open-loop (blatant rip of CheesyDrive)
and a closed-loop (spline-following via a Nonlinear Feedback
controller).

There are no unit tests for a reason; unit testing a drivetrain without
(or even with) a model is too big a hassle to be worth it. Physical
tests will suffice.